+++ Datamigratie.Server/Features/MigrateZaak/MigrateZaakService.cs
@@ -5,6 +5,7 @@ using Datamigratie.Common.Services.OpenZaak;
 using Datamigratie.Common.Services.OpenZaak.Models;
 using Datamigratie.Server.Features.MigrateZaak.Pdf;
 using Microsoft.Extensions.Options;
+using System.Diagnostics.CodeAnalysis;
 
 namespace Datamigratie.Server.Features.MigrateZaak
 {
@@ -27,8 +28,6 @@ namespace Datamigratie.Server.Features.MigrateZaak
 
             try
             {
-                CheckIfZaakAlreadyExists();
-
                 var createZaakRequest = CreateOzZaakCreationRequest(detZaak, ozZaaktypeId);
 
                 var createdZaak = await _openZaakApiClient.CreateZaak(createZaakRequest);
@@ -54,30 +53,40 @@ namespace Datamigratie.Server.Features.MigrateZaak
 
         private static OzDocument MapToOzDocument(DetDocument item, DetDocumentVersie versie, Uri informatieObjectType)
         {
+            // Apply data transformations
+            const int MaxTitelLength = 200; // 197 + "..."
+            var titel = TruncateWithDots(item.Titel, MaxTitelLength);
+
+            const int MaxBeschijvingLength = 1000; // 997 + "..."
+            var beschrijving = TruncateWithDots(item.Beschrijving, MaxBeschijvingLength);
+
+            const int MaxIdentificatieLength = 40;
+
+            // If kenmerk is longer than 40, fail the migration
+            if (item.Kenmerk?.Length > MaxIdentificatieLength)
+            {
+                throw new InvalidDataException($"Document '{item.Titel}' migration failed: The 'kenmerk' field length ({item.Kenmerk.Length}) exceeds the maximum allowed length of {MaxIdentificatieLength} characters.");
+            }
+
             return new OzDocument
             {
                 Bestandsnaam = versie.Bestandsnaam,
                 Bronorganisatie = "999990639", // moet een valide rsin zijn,
                 Formaat = versie.Mimetype,
-                Identificatie = item.Kenmerk ?? "kenmerk onbekend",
+                Identificatie = item.Kenmerk,
                 Informatieobjecttype = informatieObjectType,
                 Taal = "dut",
-                Titel = item.Titel,
+                Titel = titel,
                 Vertrouwelijkheidaanduiding = VertrouwelijkheidsAanduiding.openbaar,
                 Bestandsomvang = versie.Documentgrootte,
-                Auteur = "auteur",
-                Beschrijving = "beschrijving",
+                Auteur = "onbekend",
+                Beschrijving = beschrijving,
                 Creatiedatum = versie.Creatiedatum,
                 Status = DocumentStatus.in_bewerking,
                 Trefwoorden = [],
-                Verschijningsvorm = "verschijningsvorm",
+                Verschijningsvorm = item?.DocumentVorm?.Naam,
                 Link = ""
             };
-        }
-
-        private static void CheckIfZaakAlreadyExists()
-        {
-            // TODO -> story DATA-48
         }
         private async Task CreateAndLinkDocumentAsync(
             OzDocument ozDocument, 
@@ -231,7 +240,6 @@ namespace Datamigratie.Server.Features.MigrateZaak
             var omschrijving = TruncateWithDots(detZaak.Omschrijving, MaxOmschrijvingLength);
 
             // Now create the request
-
             var createRequest = new CreateOzZaakRequest
             {
                 Identificatie = detZaak.FunctioneleIdentificatie,
@@ -254,11 +262,12 @@ namespace Datamigratie.Server.Features.MigrateZaak
         /// <summary>
         /// Truncates the string when the length of the input string exceeds the maxLength
         /// If this happen three dots are added to the end to indiciate that the orignal value was truncated
-        /// 
+        ///
         /// The maxLength param will be the length of the string with dots
         /// Example: input: [hello world], maxlength[5] -> output: he... [length=5]
         /// </summary>
-        private static string TruncateWithDots(string input, int maxLength)
+        [return: NotNullIfNotNull(nameof(input))]
+        private static string? TruncateWithDots(string? input, int maxLength)
         {
             if (string.IsNullOrWhiteSpace(input) || input.Length <= maxLength)
                 return input;
@@ -273,7 +282,7 @@ namespace Datamigratie.Server.Features.MigrateZaak
                 return dots;
             }
 
-            var truncatedInput = input.Substring(0, maxLength - dots.Length).TrimEnd();
+            var truncatedInput = input[..(maxLength - dots.Length)].TrimEnd();
 
             return truncatedInput + dots;
         }