+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [Migrate attachments from e-Suite to OpenZaak)](https://dimpact.atlassian.net/browse/DATA-96)
 - [Migrate all versions of a Document from the e-Suite to OpenZaak](https://dimpact.atlassian.net/browse/DATA-119)
 - [Store e-Suite data as document in OpenZaak](https://dimpact.atlassian.net/browse/DATA-102)
+- [Migration of non-configurable additional Document properties from the e-Suite to OpenZaak)](https://dimpact.atlassian.net/browse/DATA-112)